pyproject.toml
[tool.poetry]
name = "ReadListManager"
version = "1.0.1"
description = "This is an ReadListManager system built using Python, Django for reading list management feature."
authors = ["Suman Sunuwar <jaycessunuwar@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
django = { version = "^4.2", extras = ["bcrypt"] }
psycopg = "^3.2.1"
django-filter = "^24.3"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.2"
flake8 = "^7.1.1"
python-dotenv = "^1.0.1"
six = "^1.16.0"
pillow = "^10.4.0"

[tool.poetry.group.dev.dependencies]
colorlog = "^6.8.2"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
pytest-cov = "^5.0.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
testpaths = ["src"]
addopts = "--cov=src --cov-report=term --reuse-db"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
